Nav, header, footer and portfolio.class files
The header, nav, and footer PHP files are included in every page of this
website via the PHP include functions that appear in the source code of
each page. These files should be modified when editing the header,
footer, or nav section of this site.

  * header.php, contains code that you would normally find in a site's
  header such as meta tags, link relation attributes, and javascript
  blocks.

  * nav.php, contains code that produces bootstrap's navbar component.
  This component acts as the main navigation bar for this site.

  * footer.php, contains HTML and CSS that creates a thin light-colored
  hr line and contains the text you would normally find in a footer such
  as copyright notice. It also contains javascript blocks to load js
  libraries that this site depends on. Script blocks are purposely
  arranged the way they are to reduce page load time.

  * portfolio.class, contains various functions that this website uses.
  At the time of writing this commit msg portfolio.class mostly contains
  functions that set various HTTP headers such as HSTS.

To avoid the content in these files from being indexed by Google and
other search engines, this include folder lives outside of the
www folder. The www folder should be set as the document root folder
in an HTTP daemon such as NGINX.